Conflicts:
	_install/alters.sql
	seotoaster_core/application/app/Widgets/Featured/views/page.phtml
	seotoaster_core/application/controllers/Backend/PageController.php
	system/js/internal/page.min.js

+++ _install/installer/resourses/seotoaster.sql
@@ -219,7 +219,7 @@ CREATE TABLE `page` (
   `last_update` timestamp NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
   `is_404page` enum('0','1') COLLATE utf8_unicode_ci DEFAULT '0',
   `show_in_menu` enum('0','1','2') COLLATE utf8_unicode_ci DEFAULT '0',
-  `order` tinyint(3) unsigned DEFAULT NULL,
+  `order` int(10) unsigned DEFAULT NULL,
   `weight` tinyint(3) unsigned DEFAULT '0',
   `silo_id` int(10) unsigned DEFAULT NULL,
   `targeted_key_phrase` varchar(255) COLLATE utf8_unicode_ci DEFAULT NULL,
